Cluster-based permutation (fast/slow trait/state)
This script compares fast and slow data, for trait and state-based approaches. Data is the mu and beta band powers, time-locked to the go cue. We used cluster-based permutation with Welch's t-test and FDR-correction on a balanced subset of the original dataset to ensure all the subjects contributed an equal number of trials to the analysis.