Cluster-based permutation test fast/slow EMG
Compare the EMG time-frequency of single trials for the fast and slow subjects and trials to get the significance difference between the two groups. This is for EMG data time-locked to the go cue. We are using the t-test and FDR on the subset of the original dataset, with equalized trials per subject.